requirements.txt
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-openai==0.2.11
langchain-text-splitters==0.3.2
langsmith==0.1.147
openai==1.56.1
pypdf==4.0.1
python-dotenv==1.0.0
chromadb==0.5.21
tiktoken==0.8.0
